REFACTOR: refactor paths (core/internal/core -> core/internal/app)
