+++ core/internal/app/auth/auth_service.go
