+++ core/internal/app/environment/environment_service.go
@@ -1,4 +1,4 @@
-package segment
+package environment
 
 import (
 	"context"
@@ -7,7 +7,7 @@ import (
 	"core/pkg/patch"
 	rsc "core/internal/pkg/resource"
 	res "core/pkg/response"
-	"core/internal/core/auth"
+	"core/internal/app/auth"
 
 	"github.com/lib/pq"
 )
@@ -19,7 +19,7 @@ func List(
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
 ) (*res.Success, *res.Errors) {
-	var o []Segment
+	var o []Environment
 	var e res.Errors
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -33,21 +33,21 @@ func List(
 
 	rows, err := db.Pool.Query(ctx, `
   SELECT
-    s.id, s.key, s.name, s.description, s.tags
+    e.id, e.key, e.name, e.description, e.tags
   FROM
-    workspace w, project p, segment s
+    workspace w, project p, environment e
   WHERE
     w.key = $1 AND
     p.key = $2 AND
     p.workspace_id = w.id AND
-    s.project_id = p.id
+    e.project_id = p.id
   `, workspaceKey, projectKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 	}
 
 	for rows.Next() {
-		var _o Segment
+		var _o Environment
 		if err = rows.Scan(
 			&_o.ID,
 			&_o.Key,
@@ -64,21 +64,21 @@ func List(
 }
 
 // Create creates a new resource instance given the resource instance
-// (*) atk: access_type <= user
+// (*) atk: access_type <= admin
 func Create(
 	atk rsc.Token,
-	i Segment,
+	i Environment,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
 ) (*res.Success, *res.Errors) {
-	var o Segment
+	var o Environment
 	var e res.Errors
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
 	// authorize operation
-	if err := auth.Authorize(atk, rsc.AccessUser); err != nil {
+	if err := auth.Authorize(atk, rsc.AccessAdmin); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		cancel()
 	}
@@ -86,7 +86,7 @@ func Create(
 	// create root user
 	row := db.Pool.QueryRow(ctx, `
   INSERT INTO
-    segment(key, name, description, tags, project_id)
+    environment(key, name, description, tags, project_id)
   VALUES
     ($1, $2, $3, $4, (
         SELECT
@@ -120,7 +120,7 @@ func Create(
 
 	// Add policy for requesting user, after resource creation
 	if e.IsEmpty() {
-		err := auth.AddPolicy(atk, o.ID, rsc.Segment, rsc.AccessAdmin)
+		err := auth.AddPolicy(atk, o.ID, rsc.Environment, rsc.AccessAdmin)
 		if err != nil {
 			e.Append(cons.ErrorAuth, err.Error())
 		}
@@ -135,11 +135,11 @@ func Get(
 	atk rsc.Token,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	segmentKey rsc.Key,
+	environmentKey rsc.Key,
 ) (*res.Success, *res.Errors) {
 	var e res.Errors
 
-	o, err := getResource(workspaceKey, projectKey, segmentKey)
+	r, err := getResource(workspaceKey, projectKey, environmentKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 	}
@@ -147,14 +147,14 @@ func Get(
 	// authorize operation
 	if err := auth.Enforce(
 		atk,
-		o.ID,
-		rsc.Segment,
+		r.ID,
+		rsc.Environment,
 		rsc.AccessService,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
-	return &res.Success{Data: o}, &e
+	return &res.Success{Data: r}, &e
 }
 
 // Update updates resource instance given an atk, key & patch object
@@ -164,16 +164,16 @@ func Update(
 	patchDoc patch.Patch,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	segmentKey rsc.Key,
+	environmentKey rsc.Key,
 ) (*res.Success, *res.Errors) {
-	var o Segment
+	var o Environment
 	var e res.Errors
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
 	// get original document
-	r, err := getResource(workspaceKey, projectKey, segmentKey)
+	r, err := getResource(workspaceKey, projectKey, environmentKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 		cancel()
@@ -183,7 +183,7 @@ func Update(
 	if err := auth.Enforce(
 		atk,
 		r.ID,
-		rsc.Segment,
+		rsc.Environment,
 		rsc.AccessUser,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
@@ -199,7 +199,7 @@ func Update(
 	// update original with patched document
 	if _, err := db.Pool.Exec(ctx, `
   UPDATE
-    segment
+    environment
   SET
     key = $2, name = $3, description = $4, tags = $5
   WHERE
@@ -222,7 +222,7 @@ func Delete(
 	atk rsc.Token,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	segmentKey rsc.Key,
+	environmentKey rsc.Key,
 ) *res.Errors {
 	var e res.Errors
 
@@ -230,7 +230,7 @@ func Delete(
 	defer cancel()
 
 	// get original document
-	r, err := getResource(workspaceKey, projectKey, segmentKey)
+	r, err := getResource(workspaceKey, projectKey, environmentKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 		cancel()
@@ -240,7 +240,7 @@ func Delete(
 	if err := auth.Enforce(
 		atk,
 		r.ID,
-		rsc.Segment,
+		rsc.Environment,
 		rsc.AccessAdmin,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
@@ -249,7 +249,7 @@ func Delete(
 
 	if _, err := db.Pool.Exec(ctx, `
   DELETE FROM
-    segment
+    environment
   WHERE
     id = $1
   `,
