+++ core/internal/app/environment/environment_util.go
