+++ core/internal/app/flag/flag_service.go
@@ -1,4 +1,4 @@
-package identity
+package flag
 
 import (
 	"context"
@@ -7,7 +7,9 @@ import (
 	"core/pkg/patch"
 	rsc "core/internal/pkg/resource"
 	res "core/pkg/response"
-	"core/internal/core/auth"
+	"core/internal/app/auth"
+
+	"github.com/lib/pq"
 )
 
 // List returns a list of resource instances
@@ -16,9 +18,8 @@ func List(
 	atk rsc.Token,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	environmentKey rsc.Key,
 ) (*res.Success, *res.Errors) {
-	var o []Identity
+	var o []Flag
 	var e res.Errors
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -32,29 +33,27 @@ func List(
 
 	rows, err := db.Pool.Query(ctx, `
   SELECT
-    i.id, i.key
+    f.id, f.key, f.name, f.description, f.tags
   FROM
-    workspace w,
-    project p,
-    environment e,
-    identity i
+    workspace w, project p, flag f
   WHERE
     w.key = $1 AND
     p.key = $2 AND
-    e.key = $3 AND
     p.workspace_id = w.id AND
-    e.project_id = p.id AND
-    i.environment_key = e.id
-  `, workspaceKey, projectKey, environmentKey)
+    f.project_id = p.id
+  `, workspaceKey, projectKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 	}
 
 	for rows.Next() {
-		var _o Identity
+		var _o Flag
 		if err = rows.Scan(
 			&_o.ID,
 			&_o.Key,
+			&_o.Name,
+			&_o.Description,
+			&_o.Tags,
 		); err != nil {
 			e.Append(cons.ErrorNotFound, err.Error())
 		}
@@ -68,19 +67,18 @@ func List(
 // (*) atk: access_type <= admin
 func Create(
 	atk rsc.Token,
-	i Identity,
+	i Flag,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	environmentKey rsc.Key,
 ) (*res.Success, *res.Errors) {
-	var o Identity
+	var o Flag
 	var e res.Errors
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
 	// authorize operation
-	if err := auth.Authorize(atk, rsc.AccessAdmin); err != nil {
+	if err := auth.Authorize(atk, rsc.AccessUser); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		cancel()
 	}
@@ -88,38 +86,41 @@ func Create(
 	// create root user
 	row := db.Pool.QueryRow(ctx, `
   INSERT INTO
-    identity(key, environment_id)
+    flag(key, name, description, tags, project_id)
   VALUES
-    ($1, $2, (
+    ($1, $2, $3, $4, (
         SELECT
-          e.id
+          p.id
         FROM
-          workspace w, project p, environment e
+          workspace w, project p
         WHERE
-          w.key = $3 AND
-          p.key = $4 AND
-          p.key = $5 AND
-          p.workspace_id = w.id AND
-          e.project_id = p.id
+          w.key = $5 AND
+          p.key = $6 AND
+          p.workspace_id = w.id
       )
     )
   RETURNING
-    id, key;`,
+    id, key, name, description, tags;`,
 		i.Key,
+		i.Name,
+		i.Description,
+		pq.Array(i.Tags),
 		workspaceKey,
 		projectKey,
-		environmentKey,
 	)
 	if err := row.Scan(
 		&o.ID,
 		&o.Key,
+		&o.Name,
+		&o.Description,
+		&o.Tags,
 	); err != nil {
 		e.Append(cons.ErrorInput, err.Error())
 	}
 
 	// Add policy for requesting user, after resource creation
 	if e.IsEmpty() {
-		err := auth.AddPolicy(atk, o.ID, rsc.Identity, rsc.AccessAdmin)
+		err := auth.AddPolicy(atk, o.ID, rsc.Flag, rsc.AccessAdmin)
 		if err != nil {
 			e.Append(cons.ErrorAuth, err.Error())
 		}
@@ -134,17 +135,11 @@ func Get(
 	atk rsc.Token,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	environmentKey rsc.Key,
-	identityKey rsc.Key,
+	flagKey rsc.Key,
 ) (*res.Success, *res.Errors) {
 	var e res.Errors
 
-	r, err := getResource(
-		workspaceKey,
-		projectKey,
-		environmentKey,
-		identityKey,
-	)
+	o, err := getResource(workspaceKey, projectKey, flagKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 	}
@@ -152,14 +147,14 @@ func Get(
 	// authorize operation
 	if err := auth.Enforce(
 		atk,
-		r.ID,
-		rsc.Identity,
+		o.ID,
+		rsc.Flag,
 		rsc.AccessService,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
-	return &res.Success{Data: r}, &e
+	return &res.Success{Data: o}, &e
 }
 
 // Update updates resource instance given an atk, key & patch object
@@ -169,22 +164,16 @@ func Update(
 	patchDoc patch.Patch,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	environmentKey rsc.Key,
-	identityKey rsc.Key,
+	flagKey rsc.Key,
 ) (*res.Success, *res.Errors) {
-	var o Identity
+	var o Flag
 	var e res.Errors
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
 	// get original document
-	r, err := getResource(
-		workspaceKey,
-		projectKey,
-		environmentKey,
-		identityKey,
-	)
+	r, err := getResource(workspaceKey, projectKey, flagKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 		cancel()
@@ -194,7 +183,7 @@ func Update(
 	if err := auth.Enforce(
 		atk,
 		r.ID,
-		rsc.Identity,
+		rsc.Flag,
 		rsc.AccessUser,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
@@ -210,13 +199,16 @@ func Update(
 	// update original with patched document
 	if _, err := db.Pool.Exec(ctx, `
   UPDATE
-    identity
+    flag
   SET
-    key = $2
+    key = $2, name = $3, description = $4, tags = $5
   WHERE
     id = $1`,
 		r.ID.String(),
 		o.Key,
+		o.Name,
+		o.Description,
+		pq.Array(o.Tags),
 	); err != nil && err != context.Canceled {
 		e.Append(cons.ErrorInternal, err.Error())
 	}
@@ -230,8 +222,7 @@ func Delete(
 	atk rsc.Token,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	environmentKey rsc.Key,
-	identityKey rsc.Key,
+	flagKey rsc.Key,
 ) *res.Errors {
 	var e res.Errors
 
@@ -239,12 +230,7 @@ func Delete(
 	defer cancel()
 
 	// get original document
-	r, err := getResource(
-		workspaceKey,
-		projectKey,
-		environmentKey,
-		identityKey,
-	)
+	r, err := getResource(workspaceKey, projectKey, flagKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 		cancel()
@@ -254,7 +240,7 @@ func Delete(
 	if err := auth.Enforce(
 		atk,
 		r.ID,
-		rsc.Identity,
+		rsc.Flag,
 		rsc.AccessAdmin,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
@@ -263,7 +249,7 @@ func Delete(
 
 	if _, err := db.Pool.Exec(ctx, `
   DELETE FROM
-    identity
+    flag
   WHERE
     id = $1
   `,
