+++ core/internal/app/identity/identity_service.go
@@ -1,4 +1,4 @@
-package flag
+package identity
 
 import (
 	"context"
@@ -7,9 +7,7 @@ import (
 	"core/pkg/patch"
 	rsc "core/internal/pkg/resource"
 	res "core/pkg/response"
-	"core/internal/core/auth"
-
-	"github.com/lib/pq"
+	"core/internal/app/auth"
 )
 
 // List returns a list of resource instances
@@ -18,8 +16,9 @@ func List(
 	atk rsc.Token,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
+	environmentKey rsc.Key,
 ) (*res.Success, *res.Errors) {
-	var o []Flag
+	var o []Identity
 	var e res.Errors
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -33,27 +32,29 @@ func List(
 
 	rows, err := db.Pool.Query(ctx, `
   SELECT
-    f.id, f.key, f.name, f.description, f.tags
+    i.id, i.key
   FROM
-    workspace w, project p, flag f
+    workspace w,
+    project p,
+    environment e,
+    identity i
   WHERE
     w.key = $1 AND
     p.key = $2 AND
+    e.key = $3 AND
     p.workspace_id = w.id AND
-    f.project_id = p.id
-  `, workspaceKey, projectKey)
+    e.project_id = p.id AND
+    i.environment_key = e.id
+  `, workspaceKey, projectKey, environmentKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 	}
 
 	for rows.Next() {
-		var _o Flag
+		var _o Identity
 		if err = rows.Scan(
 			&_o.ID,
 			&_o.Key,
-			&_o.Name,
-			&_o.Description,
-			&_o.Tags,
 		); err != nil {
 			e.Append(cons.ErrorNotFound, err.Error())
 		}
@@ -67,18 +68,19 @@ func List(
 // (*) atk: access_type <= admin
 func Create(
 	atk rsc.Token,
-	i Flag,
+	i Identity,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
+	environmentKey rsc.Key,
 ) (*res.Success, *res.Errors) {
-	var o Flag
+	var o Identity
 	var e res.Errors
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
 	// authorize operation
-	if err := auth.Authorize(atk, rsc.AccessUser); err != nil {
+	if err := auth.Authorize(atk, rsc.AccessAdmin); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		cancel()
 	}
@@ -86,41 +88,38 @@ func Create(
 	// create root user
 	row := db.Pool.QueryRow(ctx, `
   INSERT INTO
-    flag(key, name, description, tags, project_id)
+    identity(key, environment_id)
   VALUES
-    ($1, $2, $3, $4, (
+    ($1, $2, (
         SELECT
-          p.id
+          e.id
         FROM
-          workspace w, project p
+          workspace w, project p, environment e
         WHERE
-          w.key = $5 AND
-          p.key = $6 AND
-          p.workspace_id = w.id
+          w.key = $3 AND
+          p.key = $4 AND
+          p.key = $5 AND
+          p.workspace_id = w.id AND
+          e.project_id = p.id
       )
     )
   RETURNING
-    id, key, name, description, tags;`,
+    id, key;`,
 		i.Key,
-		i.Name,
-		i.Description,
-		pq.Array(i.Tags),
 		workspaceKey,
 		projectKey,
+		environmentKey,
 	)
 	if err := row.Scan(
 		&o.ID,
 		&o.Key,
-		&o.Name,
-		&o.Description,
-		&o.Tags,
 	); err != nil {
 		e.Append(cons.ErrorInput, err.Error())
 	}
 
 	// Add policy for requesting user, after resource creation
 	if e.IsEmpty() {
-		err := auth.AddPolicy(atk, o.ID, rsc.Flag, rsc.AccessAdmin)
+		err := auth.AddPolicy(atk, o.ID, rsc.Identity, rsc.AccessAdmin)
 		if err != nil {
 			e.Append(cons.ErrorAuth, err.Error())
 		}
@@ -135,11 +134,17 @@ func Get(
 	atk rsc.Token,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	flagKey rsc.Key,
+	environmentKey rsc.Key,
+	identityKey rsc.Key,
 ) (*res.Success, *res.Errors) {
 	var e res.Errors
 
-	o, err := getResource(workspaceKey, projectKey, flagKey)
+	r, err := getResource(
+		workspaceKey,
+		projectKey,
+		environmentKey,
+		identityKey,
+	)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 	}
@@ -147,14 +152,14 @@ func Get(
 	// authorize operation
 	if err := auth.Enforce(
 		atk,
-		o.ID,
-		rsc.Flag,
+		r.ID,
+		rsc.Identity,
 		rsc.AccessService,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 	}
 
-	return &res.Success{Data: o}, &e
+	return &res.Success{Data: r}, &e
 }
 
 // Update updates resource instance given an atk, key & patch object
@@ -164,16 +169,22 @@ func Update(
 	patchDoc patch.Patch,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	flagKey rsc.Key,
+	environmentKey rsc.Key,
+	identityKey rsc.Key,
 ) (*res.Success, *res.Errors) {
-	var o Flag
+	var o Identity
 	var e res.Errors
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
 	// get original document
-	r, err := getResource(workspaceKey, projectKey, flagKey)
+	r, err := getResource(
+		workspaceKey,
+		projectKey,
+		environmentKey,
+		identityKey,
+	)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 		cancel()
@@ -183,7 +194,7 @@ func Update(
 	if err := auth.Enforce(
 		atk,
 		r.ID,
-		rsc.Flag,
+		rsc.Identity,
 		rsc.AccessUser,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
@@ -199,16 +210,13 @@ func Update(
 	// update original with patched document
 	if _, err := db.Pool.Exec(ctx, `
   UPDATE
-    flag
+    identity
   SET
-    key = $2, name = $3, description = $4, tags = $5
+    key = $2
   WHERE
     id = $1`,
 		r.ID.String(),
 		o.Key,
-		o.Name,
-		o.Description,
-		pq.Array(o.Tags),
 	); err != nil && err != context.Canceled {
 		e.Append(cons.ErrorInternal, err.Error())
 	}
@@ -222,7 +230,8 @@ func Delete(
 	atk rsc.Token,
 	workspaceKey rsc.Key,
 	projectKey rsc.Key,
-	flagKey rsc.Key,
+	environmentKey rsc.Key,
+	identityKey rsc.Key,
 ) *res.Errors {
 	var e res.Errors
 
@@ -230,7 +239,12 @@ func Delete(
 	defer cancel()
 
 	// get original document
-	r, err := getResource(workspaceKey, projectKey, flagKey)
+	r, err := getResource(
+		workspaceKey,
+		projectKey,
+		environmentKey,
+		identityKey,
+	)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 		cancel()
@@ -240,7 +254,7 @@ func Delete(
 	if err := auth.Enforce(
 		atk,
 		r.ID,
-		rsc.Flag,
+		rsc.Identity,
 		rsc.AccessAdmin,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
@@ -249,7 +263,7 @@ func Delete(
 
 	if _, err := db.Pool.Exec(ctx, `
   DELETE FROM
-    flag
+    identity
   WHERE
     id = $1
   `,
