+++ core/internal/app/project/project_util.go
