+++ core/internal/app/segment/segment.go
