+++ core/internal/app/segment/segment_api.go
