+++ core/internal/app/segment/segment_util.go
