+++ core/internal/app/segmentrule/segmentrule.go
