+++ core/internal/app/segmentrule/segmentrule_api.go
