+++ core/internal/app/segmentrule/segmentrule_util.go
