+++ core/internal/app/trait/trait.go
