+++ core/internal/app/trait/trait_util.go
