+++ core/internal/app/variation/variation.go
