+++ core/internal/app/variation/variation_util.go
