+++ core/internal/app/workspace/workspace_service.go
@@ -1,4 +1,4 @@
-package project
+package workspace
 
 import (
 	"context"
@@ -7,43 +7,37 @@ import (
 	"core/pkg/patch"
 	rsc "core/internal/pkg/resource"
 	res "core/pkg/response"
-	"core/internal/core/auth"
+	"core/internal/app/auth"
 
 	"github.com/lib/pq"
 )
 
 // List returns a list of resource instances
-// (*) atk: access_type <= service
-func List(
-	atk rsc.Token,
-	workspaceKey rsc.Key,
-) (*res.Success, *res.Errors) {
-	var o []Project
+// (*) atk: access_type <= root
+func List(atk rsc.Token) (*res.Success, *res.Errors) {
+	var o []Workspace
 	var e res.Errors
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
 	// authorize operation
-	if err := auth.Authorize(atk, rsc.AccessService); err != nil {
+	if err := auth.Authorize(atk, rsc.AccessRoot); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		cancel()
 	}
 
 	rows, err := db.Pool.Query(ctx, `
   SELECT
-    p.id, p.key, p.name, p.description, p.tags
+    id, key, name, description, tags
   FROM
-    workspace w, project p
-  WHERE
-    w.key = $1 AND
-    p.workspace_id = w.id
-  `, workspaceKey)
+    workspace
+  `)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 	}
 
 	for rows.Next() {
-		var _o Project
+		var _o Workspace
 		if err = rows.Scan(
 			&_o.ID,
 			&_o.Key,
@@ -60,19 +54,15 @@ func List(
 }
 
 // Create creates a new resource instance given the resource instance
-// (*) atk: access_type <= admin
-func Create(
-	atk rsc.Token,
-	i Project,
-	workspaceKey rsc.Key,
-) (*res.Success, *res.Errors) {
-	var o Project
+// (*) atk: access_type <= root
+func Create(atk rsc.Token, i Workspace) (*res.Success, *res.Errors) {
+	var o Workspace
 	var e res.Errors
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
 	// authorize operation
-	if err := auth.Authorize(atk, rsc.AccessAdmin); err != nil {
+	if err := auth.Authorize(atk, rsc.AccessRoot); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
 		cancel()
 	}
@@ -80,20 +70,15 @@ func Create(
 	// create root user
 	row := db.Pool.QueryRow(ctx, `
   INSERT INTO
-    project(key, name, description, tags, workspace_id)
+    workspace(key, name, description, tags)
   VALUES
-    ($1, $2, $3, $4, (
-      SELECT id
-      FROM workspace
-      WHERE key = $5
-    ))
+    ($1, $2, $3, $4)
   RETURNING
     id, key, name, description, tags;`,
 		i.Key,
 		i.Name,
 		i.Description,
 		pq.Array(i.Tags),
-		workspaceKey,
 	)
 	if err := row.Scan(
 		&o.ID,
@@ -107,7 +92,7 @@ func Create(
 
 	// Add policy for requesting user, after resource creation
 	if e.IsEmpty() {
-		err := auth.AddPolicy(atk, o.ID, rsc.Project, rsc.AccessAdmin)
+		err := auth.AddPolicy(atk, o.ID, rsc.Workspace, rsc.AccessAdmin)
 		if err != nil {
 			e.Append(cons.ErrorAuth, err.Error())
 		}
@@ -116,16 +101,12 @@ func Create(
 	return &res.Success{Data: o}, &e
 }
 
-// Get gets a resource instance given an atk & key
+// Get gets a resource instance given an atk & workspaceKey
 // (*) atk: access_type <= service
-func Get(
-	atk rsc.Token,
-	workspaceKey rsc.Key,
-	projectKey rsc.Key,
-) (*res.Success, *res.Errors) {
+func Get(atk rsc.Token, workspaceKey rsc.Key) (*res.Success, *res.Errors) {
 	var e res.Errors
 
-	r, err := getResource(workspaceKey, projectKey)
+	r, err := getResource(workspaceKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 	}
@@ -134,7 +115,7 @@ func Get(
 	if err := auth.Enforce(
 		atk,
 		r.ID,
-		rsc.Project,
+		rsc.Workspace,
 		rsc.AccessService,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
@@ -143,21 +124,20 @@ func Get(
 	return &res.Success{Data: r}, &e
 }
 
-// Update updates resource instance given an atk, key & patch object
+// Update updates resource instance given an atk, workspaceKey & patch object
 // (*) atk: access_type <= user
 func Update(
 	atk rsc.Token,
 	patchDoc patch.Patch,
 	workspaceKey rsc.Key,
-	projectKey rsc.Key,
 ) (*res.Success, *res.Errors) {
-	var o Project
+	var o Workspace
 	var e res.Errors
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
 	// get original document
-	r, err := getResource(workspaceKey, projectKey)
+	r, err := getResource(workspaceKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 		cancel()
@@ -167,7 +147,7 @@ func Update(
 	if err := auth.Enforce(
 		atk,
 		r.ID,
-		rsc.Project,
+		rsc.Workspace,
 		rsc.AccessUser,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
@@ -183,12 +163,12 @@ func Update(
 	// update original with patched document
 	if _, err := db.Pool.Exec(ctx, `
   UPDATE
-    project
+    workspace
   SET
     key = $2, name = $3, description = $4, tags = $5
   WHERE
-    id = $1`,
-		r.ID.String(),
+    key = $1`,
+		workspaceKey,
 		o.Key,
 		o.Name,
 		o.Description,
@@ -200,19 +180,15 @@ func Update(
 	return &res.Success{Data: o}, &e
 }
 
-// Delete deletes a resource instance given an atk & key
+// Delete deletes a resource instance given an atk & workspaceKey
 // (*) atk: access_type <= admin
-func Delete(
-	atk rsc.Token,
-	workspaceKey rsc.Key,
-	projectKey rsc.Key,
-) *res.Errors {
+func Delete(atk rsc.Token, workspaceKey rsc.Key) *res.Errors {
 	var e res.Errors
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
 	// get original document
-	r, err := getResource(workspaceKey, projectKey)
+	r, err := getResource(workspaceKey)
 	if err != nil {
 		e.Append(cons.ErrorNotFound, err.Error())
 		cancel()
@@ -222,7 +198,7 @@ func Delete(
 	if err := auth.Enforce(
 		atk,
 		r.ID,
-		rsc.Project,
+		rsc.Workspace,
 		rsc.AccessAdmin,
 	); err != nil {
 		e.Append(cons.ErrorAuth, err.Error())
@@ -231,16 +207,11 @@ func Delete(
 
 	if _, err := db.Pool.Exec(ctx, `
   DELETE FROM
-    project
+    workspace
   WHERE
-    workspace_id = (
-      SELECT id
-      FROM workspace
-      WHERE key = $1
-    ) AND key = $2
+    key = $1
   `,
 		workspaceKey,
-		projectKey,
 	); err != nil {
 		e.Append(cons.ErrorInternal, err.Error())
 	}
